Pancreatitis-main/Pancreatitis-main/pancreatic_cancer_prediction/frontend/requirements.txt
 streamlit
 pandas
 numpy
 scikit-learn
 joblib
 matplotlib
 seaborn
 google-generativeai
